fixed dashboard navigations
