+++ src/views/Dashboard/DashContainer/DashContainer.jsx
@@ -1,19 +1,11 @@
+import DashBoardNav from "../DashBoardNav/DashBoardNav"
 import Main from "../Main/Main"
-import SideNav from "../SideNav/SideNav"
 
 const DashContainer = () => {
     return (
-        <div className="container-fluid p-3">
-            <div>
-                <SideNav />
-            </div>
-
-
-
-
-                    <Main />
-
-
+        <div>
+            <DashBoardNav />
+            <Main />
         </div>
     )
 }
